webui: Improve the Storage Deals page
